# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/email/login": {
      "post": {
        "operationId": "AuthController_emailLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Вход по email и паролю",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/telegram/bot-login-url": {
      "get": {
        "operationId": "AuthController_botLoginUrl",
        "parameters": [
          {
            "name": "partnerSlug",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/telegram/bot-login-poll": {
      "get": {
        "operationId": "AuthController_botLoginPoll",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/payments/order/create": {
      "post": {
        "operationId": "PaymentController_createOrder",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/order/{orderId}": {
      "get": {
        "operationId": "PaymentController_getOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
...
    "/partner/apply": {
      "post": {
        "operationId": "PartnerController_apply",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
    "/partner/application": {
      "get": {
        "operationId": "PartnerController_application",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
    "/partner/dashboard": {
      "get": {
        "operationId": "PartnerController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
    "/partner/payout": {
      "post": {
        "operationId": "PartnerController_payout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
    "/partner/payout-wallet": {
      "post": {
        "operationId": "PartnerController_payoutWallet",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
    "/partner/currency": {
      "post": {
        "operationId": "PartnerController_currency",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "partner"
        ]
      }
    },
...
  },
  "info": {
    "title": "Neiro API",
    "description": "OpenAPI 3 спецификация, сгенерированная из NestJS-контроллеров (@nestjs/swagger).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Authorization: Bearer <token> (как в Express-сервере)"
      }
    },
...
  }
}